requirements-dev.txt
debugpy
flask
pytest
mock
pytest-mock
fake-bpy-module-2.93